requirements.txt
# install pytorch 1.2 from conda
cython
pre-commit
PyYAML==5.1.1
sacred==0.7.5
SQLAlchemy==1.3.5
psycopg2-binary==2.8.3
matplotlib==3.1.0
pytorch-transformers==1.1.0
colorlog==4.0.2
pytrec-eval==0.4
pycapnp==0.6.4
pytest==3.8.0
pyjnius==1.2.0
pescador==2.0.2
pymagnitude==0.1.120
h5py==2.9.0
pytorch-pretrained-bert==0.4
nltk==3.4.5
pymongo==3.9.0
# deps that the pymagnitude package isn't pulling in:
lz4==2.1.10
xxhash==1.3.0
annoy==1.15.2
fasteners==0.15
#
django==2.2.1
pytest-mock==1.10.4
mock
pyserini==0.7.0.0
scipy==1.3.0
